third_party/blink/web_tests/fast/dom/HTMLElement/insertAdjacentHTML-errors-expected.txt

Test insertAdjacentHTML exceptions to make sure they match HTML5

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS div.insertAdjacentHTML('beforeBegin', 'text') threw exception NoModificationAllowedError: Failed to execute 'insertAdjacentHTML' on 'Element': The element has no parent..
PASS div.insertAdjacentHTML('afterEnd', 'text') threw exception NoModificationAllowedError: Failed to execute 'insertAdjacentHTML' on 'Element': The element has no parent..
PASS div.insertAdjacentHTML('FOO', 'text') threw exception SyntaxError: Failed to execute 'insertAdjacentHTML' on 'Element': The value provided ('FOO') is not one of 'beforeBegin', 'afterBegin', 'beforeEnd', or 'afterEnd'..
PASS document.documentElement.insertAdjacentHTML('afterEnd', 'text') threw exception NoModificationAllowedError: Failed to execute 'insertAdjacentHTML' on 'Element': The element has no parent..
PASS successfullyParsed is true

TEST COMPLETE
